chore: consolidate existing hooks into /hooks
